Refactor SCRAM-SHA-1 implementation into an abstract base class
Extract common SCRAM protocol handling logic from the SCRAM-SHA-1 implementation into a new abstract ScramSaslServer base class.

This refactoring centralizes:
- SCRAM state handling and message flow
- GS2 header parsing
- channel binding validation
- downgrade attack protection
- nonce generation and verification
- client/server proof verification
- fake credential generation for timing attack mitigation
- fake salt generation for user enumeration protection
- SASL lifecycle handling

Mechanism-specific implementations now provide only:
- hashing/HMAC primitives
- credential retrieval
- iteration/salt configuration
- algorithm-specific parameters

This significantly reduces duplication between SCRAM variants and establishes a reusable foundation for additional SCRAM mechanisms such as SCRAM-SHA-256 and SCRAM-SHA-256-PLUS.
